requirements.txt
python-dotenv
streamlit
pdf2image
google-generativeai
Pillow
deep_translator